backend/neo4j_knowledge_graph_test/Coastline KG/requirements.txt
graphiti-core[falkordb]>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
